add profile id to all mappings

+++ include/config.hpp
@@ -31,7 +31,7 @@ extern bool working;
 extern std::vector<std::shared_ptr<Instance>> instances;
 extern std::map<uint8_t, std::shared_ptr<UsbDevice>> usb_instances;
 extern std::map<uint32_t, std::shared_ptr<Device>> devices;
-extern std::map<uint8_t, std::shared_ptr<Instance>> active_instances;
+extern std::map<uint32_t, std::shared_ptr<Instance>> active_instances;
 inline bool hid_based(void)
 {
     return mode == ModeHid || mode == ModePs3 || mode == ModePs4 || mode == ModeWiiRb || mode == ModeSwitch;
